[BUILD] docker-compose.yml: Docker environment setup and configuration modifications
The .dockerignore file was created to specify intentionally untracked files. The Dockerfile was updated to correct the application output name from `runecraftd` to `pusgOff` and the command to run the application from `./main.go` to `./pushOff`. Moreover, a `docker-compose.yml` file was introduced to define services, including `push_off` and `db`.