[BUILD] Dockerfile: Build Dockerfile changes
The Dockerfile was modified to fix errors that prevented correct compilation. Specifically, the `go build` command was updated to change the output binary from `pusgOff` to `pushOff`, and the `COPY` command was updated to copy the correct binary file. Additionally, the version of `openssh-client` was changed from `10.2p1` to `10.0_p1-r9`.

The `WORKDIR` was set to `/root/` and the base image remains at `alpine:3.22.2`, while package installation adjustments were made.