[ADD] main.go: Adding whitelist for service access
The `main.go` file was modified to enable the use of a whitelist of users who can access the service. Specifically, the lines related to setting up host key paths and authorized keys were updated to use `wish.WithHostKeyPath(hostKeyPath)` and `wish.WithAuthorizedKeys(authorizedKeysPath)`, respectively. This change allows for specifying the path to SSH keys, which is set to be the same as the volume of the Docker container.